Added remove function
Added a function to remove items from the checkout list by clicking on
them.  After some debugging to see why it was not working, found the
part that was causing the problems and replaced it with a string that
contains inline html code that is added with the buttons, by a function.