Fixed bug with Reader function, and ReadRow method.
The ReadRow method in JSPOS.js did not have a bug itself, causing it to
return all the items of the menu table.  It was a problem with using the
Reader function in Try.php.

The result of calling Reader function needed to be returned, that is why
I had that bug.