Fixed subtraction bug with rg.adjust_price
The value that was taken off by rg.adjust_price, was not put back on
when the price adjustment was undone.    It was subtracting a positive
float to apply the adjustment, and was subtracting a positive float when
undoing it!     It  was fixed by changing the initial subtracting to an
addition of a negative float.