Fixed bug with item duplication
There was a bug such that when the quantity button would be clicked, it
would duplicate the selected item a certain number of times, but the
subtotal would be the same as if it was only one item.

This was caused by the statement that updates the value of rg.subtotal
being out side the while loop in rg.ring_up().
It is that loop that duplicates items if rg.dup > 1.